Change NVIM_HOME_PROFILE check to NVIM_PROFILE == "home"
Apparently environment variables are always loaded as strings, so I
can't check for booleans immediately. This makes it more convenient to
just assign a string to the variable at the shell level instead of
making a boolean environment variable for each different profile (in
case more profiles are considered in the future).